+++ expo-horizon-location/android/src/quest/java/expo/modules/LocationHelpers.kt
@@ -198,7 +198,7 @@ class LocationHelpers {
               continuation.resume(PermissionRequestResponse(result))
             }
 
-            override fun reject(code: String, message: String?, cause: Throwable?) {
+            override fun reject(code: String?, message: String?, cause: Throwable?) {
               continuation.resumeWithException(CodedException(code, message, cause))
             }
           },
@@ -220,7 +220,7 @@ class LocationHelpers {
               )
             }
 
-            override fun reject(code: String, message: String?, cause: Throwable?) {
+            override fun reject(code: String?, message: String?, cause: Throwable?) {
               it.resumeWithException(CodedException(code, message, cause))
             }
           },